# 003 Code/Aiary-Backend-Public-main/swagger-api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/chat/messages": {
      "post": {
        "tags": [
          "Chat Message"
        ],
        "summary": "사용자 메시지 전송",
        "description": "스레드에 사용자 메시지를 추가합니다",
        "operationId": "postMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "스레드를 찾을 수 없음",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "409": {
            "description": "스레드가 finalized 상태",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "메시지 전송 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/chat/messages/{messageId}": {
      "delete": {
        "tags": [
          "Chat Message"
        ],
        "summary": "메시지 삭제",
        "description": "특정 메시지를 삭제합니다",
        "operationId": "deleteMessage",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "메시지 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "메시지 삭제 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "메시지를 찾을 수 없음",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "스레드가 finalized 상태",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Chat Message"
        ],
        "summary": "메시지 수정",
        "description": "특정 메시지의 내용을 수정합니다",
        "operationId": "patchMessage",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "메시지 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "메시지를 찾을 수 없음",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "스레드가 finalized 상태",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "메시지 수정 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/threads/{threadId}/messages": {
      "get": {
        "tags": [
          "Chat Message"
        ],
        "summary": "메시지 목록 조회",
        "description": "특정 스레드의 메시지 목록을 페이지네이션으로 조회합니다",
        "operationId": "listMessages",
        "parameters": [
          {
            "name": "threadId",
            "in": "path",
            "description": "스레드 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "조회할 항목 수 (최대 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            },
            "example": 50
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "페이지네이션 커서 (이전 응답의 nextCursor 값)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "메시지 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/chat/stream": {
      "get": {
        "tags": [
          "Chat Message"
        ],
        "summary": "AI 응답 스트리밍",
        "description": "스레드의 메시지를 기반으로 AI 응답을 실시간 스트리밍합니다",
        "operationId": "stream",
        "parameters": [
          {
            "name": "threadId",
            "in": "query",
            "description": "스레드 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "스레드를 찾을 수 없음",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SseEmitter"
                }
              }
            }
          },
          "409": {
            "description": "스레드가 finalized 상태",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SseEmitter"
                }
              }
            }
          },
          "200": {
            "description": "스트리밍 시작 성공",
            "content": {
              "text/event-stream": {}
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SendMessageRequest": {
        "type": "object",
        "properties": {
          "threadId": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          }
        }
      },
...
      "SseEmitter": {
        "type": "object",
        "properties": {
          "timeout": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    }
  }
}